[Inspector][Rules] Ensure pseudo element rules are expanded in checkAppliedRules tests
In the React-based Rules view, the pseudo element rules are not rendered unless it is expanded
compare to the old Rules view which are rendered and hidden.

We ensure pseudo element rules are rendered by expanding them in the the test harness prior
to checking for the right applied rules.