[Inspector] Part 6: Push the angle, timing-function, flex and grid values into the parsed array
In the future, the OutputParser will only return an array of objects which can be consumed
by a renderer such as React to render the correct components for a given property value.

This patch removes the remaining DOM nodes that are created for angle, timing-function,
flex and grid. In addition, simplifies the angle, timing-function, flex and grid handling
by only pushing the text along with their type information into the parsed array that
will be returned from the OutputParser.