mixpanel: track breakpoints
Use mixpanel to track who sets breakpoints, in prod only.

Also add userEmail to mixpanel and sentry context.

We're just using mixpanel as an experiment right now, it might be
removed in the future. We want to see how useful it is to capture
user behavior that we can't just infer from rows written to the DB.