Reduce unmounts when hiding DevTool panes
This started as an investigation into
https://github.com/RecordReplay/devtools/issues/3851. Toggling the
visibility of the Editor was triggering an unmount and remount of
basically every component in the DevView. DevView had 4 different
branches, each with *slightly* different rendering trees, based on which
components (editor, video, both, neither) were open. This was a lot of
complexity *and* it actually didn't capture all of the options (because
the user can also toggle whether or not the sidebar was open).

I cleaned it up by simplifying that render tree so that it's always the
same child components inside of the `SplitBox`. There is still one place
where changing a view causes an unmount and remount which is inside of
`Viewer`. This is because in the vertical view the video is the
`startComponent` (because vertical SplitBoxes flow top-to-bottom), but
in the horizontal view the video is supposed to be the *endComponent*
(because horizontal SplitBoxes flow left-to-right). I thought I would be
clever and just add `flex-direction: row-reverse` in the horizontal case
but that caused a really horrible UI bug where trying to slide the
component to make it bigger actually made it smaller and vice versa. I
was far enough down the rabbit-hole by this point that I didn't want to
spend any more time going that route.

This change also addresses
https://github.com/RecordReplay/devtools/issues/3746. It looks like
previously viewer mode stored a panel split size in
`nonDevSidePanelWidth`, but the function that sets that was no longer
called, so I've removed that setting. Instead, both places where
`SidePanel` is rendered now pull from and write to the setting
`sidePanelSize`, in the new `prefs` file. This means that when you
navigate between DevTools and Viewer mode the width of the side panel
will remain the same.

I cannot tell if this is fully intended, but it works quite well in my
testing. Consider the following case, where the container is a SplitBox
flowing left-to-right, with component A being the controlled component
(which means component B has flex-grow on it and will take up whatever
space A leaves open):

![image](https://user-images.githubusercontent.com/5903784/135940611-c07e8b12-cf82-468f-9073-7a18e8f8b965.png)

Let's also say that component A might not be visible right now, in which
case we want component B to take up its space, ignoring the partition.
Rather than doing this:

```
if (!componentA) {
 return componentB;
} else {
  return <SplitBox left={componentA} right={componentB}/>
}
```

You can do something like this:

```
return <SplitBox left={componentA} right={componentB} maxWidth={componentA ? "auto" : "0px"}/>
```

This says "give me a SplitBox that contains these two components, but if
this one doesn't exist, then make sure it doesn't take up any space".
Without code like this, it's actually pretty difficult to avoid having
many different `if` checks in the render statements or having accidental
gaps in the UI where the SplitBox is making room for an element that
doesn't exist.

This gives us the following:

https://user-images.githubusercontent.com/5903784/135941252-b0be4582-ef85-4e20-912e-0171a9a5f397.mp4

I think this is in a pretty nice state for the user interface, but I
don't love how it feels to code UI's like this with the nested
SplitBoxes. I think a future we should move towards, particularly for
layouts like DevTools, is a CSS grid, where the drag events just change
the CSS, rather than the react state, and things can be moved from panel
to panel by just changing their grid number. That's a much larger
undertaking.