Change default video viewer height to 375px
This was set to 50%, which would have been a good default, except that
in `SplitBox` we run `parseInt` on the `initialSize` prop. This turns
`50%` into just `50`, which then gets interpreted as a number of pixels.
It's actually pretty difficult to do a starting size in percentages,
because of the way we have to interpret cursor movement into pixel sizes
when resizing the SplitBox. If we're sticking with pixels, 375 is much
more reasonable, since that is the size that the video maxes out at
anyways.