Show the current time in the timeline
Closes #3822 , #3320 is coming up next.

Changes
---

- Split out some of the components in the Events folder, to try and give
  them some room to grow, since I'm about to add a bunch of logic.
- Add `lightGrey` to the colors that our tailwind knows about
- Add a secondary accent (the *now* color) line that shows where in the
  event timeline the current time is, and fade events that have not yet
  happened. This is quite pleasing to watch as the video plays.
- Fix what I *think* is a bug. Events on the timeline are represented in html as `button` elements, so that they behave like they are clickable (which makes sense, because they are). When you click on one of them, it gets focus. If you then press space bar two things happen at once:

1. The viewer attempts to start playing
2. The viewer immediately jumps to that point in time and pauses

It manifests as like a small UI hiccup every time you press the space
bar while having focus on a timeline event. I could actually see a good
argument for allowing timeline events to work this way, if people wanted
to use the keyboard to navigate the timeline, but I think most of the
time this behavior is just confusing, and if they *really* want to tab
through and select they can always hit "Enter", which has the same
effect as space, but doesn't try to start the player first.

https://user-images.githubusercontent.com/5903784/137043297-bf7d02b6-a69d-49af-907d-cabe9ccacee3.mp4