Manage focus with a React context
This is more complicated than I expected! It's complicated for these reasons:
- There are multiple editors per CommentCard (at least one, that is
  disabled by default, for the top-level comment, and at least one for
  the enabled reply box [assuming we're paused there, otherwise none of
  this really matters])
- Every part of the comment serves as a click target to focus the
  related reply box
- Reply boxes can appear and disappear (sometimes several at once!), so
  you have to think about both competing for focus as well as competing
  for autofocus.

I went with a context here. This is one of the scenarios where I've seen
them work pretty well, and it makes it much easier that we don't have to
thread focus information through like 4 layers of components.