Do not change sidepanel width when hiding editor
Since my last big layout PR (#3860) we've really standardized on keeping
the side panel the same width across all views. The one situation this
breaks in is in DevView when the editor is closed. The "controlled"
component of SplitBoxes that are displayed horizontally is the beginning
element, so in order to correctly adjust that width you need to set a
min/max accordingly. This adds one more check to our max width
calculations, so that in DevView, the Toolbox's width is:

- 0 when both the side panel and the editor are hidden
- the width of the side panel when the editor is closed
- free to be resized between the min-side-panel-size and 80% when both
  the side panel and the editor are showing.

All of this would also be considerably cleaner with grid, but I'll
refactor that another day!

Fixes #3924