Parse dates when sorting
I thought that we were doing date parsing when we loaded information
from the API, but it turns out that we are just passing around strings,
so even the more intelligent `sortBy` from `lodash` will result in
string comparison rather than real datetime comparison. I've changed the
tests to reflect this requirement as well. I've made two other small
tweaks: what we were passing back from `commentKeys` was not *quite* the
right permutation map. I'm not a combinatorics programmer so I don't
know the word for this thing (maybe it's the transposition map?), but
it's like the inverse permutation map. Running `indexOf` gets back the
map that we actually want.

Also, it's totally fine to sort by user and then comment created at, but
it's probably more accurate to think of user as the tie breaker, so I've
made that the second sort condition. This won't change any behavior, but
I think it makes it easier to verify the logic in simple cases.