protocol: make dispatch URL a build time variable
It can still be overridden with a URL param, but this paves the way for
the world that Ryan and I talked about going to earlier today where
everything is a build time variable. Here's what that would look like:

* For local dev build on-demand frontends that are configured to talk
the appropriate backend services
* For tests I'm not sure what to do, probably will involve generating
frontends that are configured to talk to the appropriate backend ahead
of time.
* For frontend folks nothing changes, they already don't typically
need to adjust these variables