Allow embedded links to be copied and pasted
There were a couple of weird issues with the regexes we were using when
they were pasted in the middle of content. This is probably a rare case
but it's nice to handle it correctly. We will now not match a malformed
Replay URL, which should help users realize that their link isn't fully
copied in the case that they try and paste an incomplete URL. We also
render text correctly, so you can paste *from* a replay comment and get
the link back. For example:

https://user-images.githubusercontent.com/5903784/139773469-4d3ae80a-b5e1-4d5a-92fc-c124257a180e.mp4

Copying a smart link from one comment and pasting it into another
comment now *also* works. The problem here was that I had set up our
`parseHTML` matcher to `a[href]`, which means that if you pasted *any*
HTML that had a valid link in it, TipTap would try and interpret that as
a GitHub link. Now, we use `a[data-github-link]` and when we
`renderHTML` we make sure that all github links have that data attribute
(and Replay links get a similar `data-replay-link` attribute). Again,
the mention extension was  indispensable in figuring out why this was
broken. I find that reading other extension code almost always answers
my TipTap questions.

Fixes #4258