Don't connect material-icons to the store
I'm all for many small connected components, rather than massive
containers that have to pass things down through 50 layers, but it seems
a little ridiculous to me that each icon is individually `connect`ed to
the redux store.

I'm not sure if the #app-container addition is the right way to do this,
but I actually found it difficult to add a class to the highest level.
If anybody has suggestions there lmk.