Allow text-like responses to be copied
Responses that register as JSON can already be copied from the JSON
viewer, but this adds copying for text responses (or content-types that
we're registered as text-like *enough* to be rendered as text, e.g.
`socket.io` sends things in `application/octet-stream`, but it's
*mostly* legible plaintext UTF-8, with a couple of preceding control
characters).

This also moves the `download response` button to a similar style.

https://user-images.githubusercontent.com/5903784/149573669-06b06cc6-4b7d-4adb-af14-b0ce3f952529.mp4