Show requests with no responses
There are times when things go wrong for an HTTP request too early in
the process for a response event to get recorded. In that case we still
want to show the request (whereas right now we filter those out,
expecting them to eventually fully load with both a request and response
event).

Changes
---
- Show requests that have no responses
- When a request is to the root of a hostname, show the name as `/`, not
  an empty field
- Move the close button a little more left, in the Replay browser it
  sometimes gets hidden by the scrollbar