Add a simple timings panel #5074
We waited on this because when I first tried to verify timings against
what Chrome and Firefox were reporting I was not seeing the same
numbers, but hitting a server where I am manually controlling the time
that the server takes to respond, everything looks good. That makes me
think there were just legitimate time differences in the requests I was
recording before.

And remove Coming Soon because all tabs are populated!
![CleanShot 2022-01-21 at 13 50 35](https://user-images.githubusercontent.com/5903784/150605141-88d90b70-9487-4af1-8bff-38b562375320.png)

Other Changes
---
- There was an errant `0` when the request had no query params
- Remove the "Coming soon!" component, because all tabs are populated 😄
- Try and standardize the left-padding that all of the tabs use (this is
  hard because some have collapsed/expanded state to show and others
  don't)