Memoize alternate source IDs #5299
I will be honest: I don't *really* understand what this code does, so I
am prepared for this be a big mistake. However, I recently profiled a
Replay that was loading slowly. It had 51,000 sources, which was clearly
the thing that was causing issues somehow. I also noticed that when
loading this Replay, my browser tab would often crash with an
out-of-memory error. Trying to *record* the thing loading was, for me,
totally impossible. At some point I had to force quit Replay while
trying to record this thing, because it was currently using 11GB of
memory!

By digging in with the profiler and with Replay (I recorded Replays with
smaller numbers of sources so that I could analyze them for order of
magnitude differences) I saw that one loop in particular was hit a *lot*
of times. It was pretty unclear to me what it was doing, but by counting
the hits during load of the broken Replay I learned that during this
loading process that loop was hit over 69 *million* times. I decided
that since there were only 51,000 sources, it was pretty likely that we
were duplicating work in there *somewhere*, so I memoized the function.
After that it was called 47,000 times, which is... roughly 1:1 with
sources?

Like I said, I did my best to understand what this code was doing, but I
never really got there, so this whole thing might be a mistake, but I
figured it was worth a shot (and I suspect that if this *is* a mistake
then @hbenl will probably set me straight pretty quickly, because I
think he's much more familiar with this code than I am).