Don't scroll CodeMirror on space press
I mostly fixed this months ago in cdaea8c but over the past few days I
noticed that I still sometimes scroll the editor when I'm trying to
toggle play/pause in the video editor. It turns out that when you click
on a breakpoint inside of a breakpoint panel the element that gets focus
is CodeMirror-scroll. I'm not sure why that thing gets focus, but if I
had to guess I would say that CodeMirror detects a click on itself and
tries to get ready to scroll (which would be a pretty reasonable
shortcut in most cases). In our case, though, it's really annoying to
scroll the editor when you instead mean to play or pause the video.