Don't reset the global filter when data changes #5824
I'm not sure why react table thinks that data is changing, because this
seems to happen even at random times when not clicking on a row. Maybe
it also happens when we receive network events from the server? Anyways,
this fixes the issue.

For a little more context:

https://react-table.tanstack.com/docs/api/useGlobalFilter

> autoResetGlobalFilter: Boolean
Defaults to true
When true, the globalFilter state will automatically reset if any of the following conditions are met:
data is changed
To disable, set to false
For more information see the FAQ ["How do I stop my table state from automatically resetting when my data changes?"](https://react-table.tanstack.com/docs/faq#how-do-i-stop-my-table-state-from-automatically-resetting-when-my-data-changes)