Remove the video DOM node ref from the Redux store
DOM nodes are not "state", and not serializable, so they shouldn't
go in the Redux store.

Here, the Redux store was being used to pass around the DOM node for
the playing video, so that `<Video>` could save the node ref, and
`graphics.ts` could later call methods on it.  Understandable as a
hack, but not a good idea.

We can accomplish the same thing by just having a tiny module with
a getter/setter pair, and store the DOM node ref there.