Fix broken filtering of console messages
In f3e973a , I made several changes to `actions/filters.ts`:

- I moved calculation of new filter state into the reducer
- I removed the `FILTER_TEXT_SET` and `FILTER_TOGGLE` action types
- I stopped passing `filtersState` into the actions

Those turned out to be critical for the `messages` reducer.
As a temporary workaround, I am specifically dispatching an action
to trigger an update to the messages state for those cases.

Long-term, the right answer is to make that all "derived" data that
is calculated in a selector, instead of having the reducer calculate
that itself.