Consolidate webconsole pref persistence logic
Previously, the "log level" prefs were being persisted in a thunk,
in `actions/filters`.  Jason wants pref persistence to be reactive
and in a single location, in `ui/setup/prefs`.

Also, the `timestampsVisible` flag wasn't being persisted properly,
and the `node_modules` pref looks like it was never persisted.  Both
of those are now persisted correctly.