Use a number for accordion initialHeight #6026
Originally I thought this was clever, but it actually fails when you go
to adjust the heights, because the reducer does not know how to deal
with percentages. The real solution here would be to make the reducer
smart enough to convert back and forth between percenages and pixel
values, but I think that's quite a bit harder than just throwing a
reasonable number in there and calling it a day.