Disable commenting, editing, etc. when updating
First off, actually return promises from our update/create hooks, so
that we can properly await them.

This will mark a comment as updating, which will disable all of the
various paths to editing it and its replies. It will also make it
translucent, to let the user know that stuff is still in progress.

Follow-up: This GraphQL operation to insert a comment record should not
take 5.5 seconds.