Remove unneeded comment selectors
We had a couple of unused connections sitting around, so nice to get
those out of there. Also, as far as I can tell, this use of
pendingComment in timeline does not do anything? I could be wrong about
this one, but I pulled it out and couldn't see any difference in
behavior.