Add a hitCount line limit
Right now we have a very naive algorithm: fetch hitCounts for all lines
when a source file gets hovered for the first time. This can swamp a
controller if a file is tens of thousands of lines long, but it's *very*
useful in the case that a file is of a reasonable length. In my
experience, 1000 line fetches at once still feels pretty instantaneous,
and I think many files that people are interested in having analyzed
completely will be under 1000 lines. The hardest part of this change was
just understanding how `memoizableAction` works for the third time 🙃

In case the above did not make sense, here's an example:

- The user loads a file
- SourceOutline sees that a new file is loaded and asks for the hitCount
  on line 1
- memoizableAction will see that the line requests was 1, and actually
  ask for the hits for lines 1-1000.
- If the user scrolls to line 1005 (for example) and hovers,
  memoizableAction will fetch the hits for lines 1000-2000, etc.

This does not totally resolve
https://github.com/RecordReplay/backend/issues/5132 and
https://github.com/RecordReplay/backend/issues/5108, but it helps a
*lot*.