Address pre-existing 'react-hooks/exhaustive-deps' rule violations. (#6458)
In most cases, I tried to _fix_ the rule violation.

When the fix seemed unclear (e.g. I wasn't sure if the change would "break" the desired behavior) then I left a `TODO [authorname]` comment for the original author. (We'll rely on Trunk hold-the-line to prevent new ones from being added.)

Things to keep in mind about the following changes:
* `useEditor` from `@tiptap/react` returns a _stable_ "editor" provided the dependencies array values don't change.
* `useDispatch` from `react-redux` returns a _stable_ `dispatch` function so long as you don't change the Redux `store` (which we don't).
* There were a few places we could have been memoizing that we weren't, so I added them (e.g. `useGetRecording()` -> `convertRecording(recording.data)`).