Only request at most one column's location per line
We might actually want to check whether the user has columnBreakpoints
turned on, because theoretically someday we could show different numbers
hovering on different parts of the line, but for right now, we always
show the first location, so let's just ask for that hitCount.

A nice consequence of this is that it is a lot less likely that we will
have to fallback to analyses because we are asking for too many
locations, because the backend's limit is 1,000 locations, we query in
1,000 line chunks, and so if no line asks for more than one location we
should always be under the limit. (This is not actually necessarily true
because generated locations also have a limit but whatever, close
enough).