Hide the hitCount tooltip on breakpoint lines
See https://www.loom.com/share/bba1f5fad9f54f95914c8747ee420e35 or
https://www.notion.so/replayio/Loading-Region-Problems-b8dc0d03c49840f8b504e967c7ea5778
for a longer format explanation of why this happens, but basically,
hitCounts and breakpoint point counts might disagree while focused. And,
since the point count will be right there anyways, we don't really need
to show the hitCount on that same line. True, there are other places for
this difference to be noticed (for instance, in a straight-forward block
on the line before this breakpoint line I would expect that the
hitCounts would match this breakpoint's point count). But at least this
is deemphasizes the "wrong" count a bit and (hopefully) reduces
cognitive dissonance for the user.