Unsquish console message badge
I noticed that if the console message is really long the badge ends up
getting squished, so I've set a minWidth in addition to a width, which
seems to work. Also, I've migrated this particular set of styles to
using numbers instead of `"Xpx"`. I've only ever seen numbers elsewhere
in this codebase  (the px is implied), so I'm just going for consistency
here.