Set a low water mark for our progress loading bar
I know we punted this to next week but I was feeling tired of working on
hard things and wanted a win. To test it, I made our progress bar
increment by a negative amount instead of a positive amount half the
time.

Here's what that looked like before this change:

https://user-images.githubusercontent.com/5903784/167541920-7de2c9f9-dfc6-4e53-bca5-9360e7806ace.mp4

And after (note, it doesn't move forward very quickly, because half the
time the progress is actually decreasing, but it doesn't move
backwards):

https://user-images.githubusercontent.com/5903784/167541872-31606ea8-7a6d-4f4d-b5d3-a7d2fd437004.mp4

Fixes https://github.com/RecordReplay/devtools/issues/6500