Show when all paints loaded
The protocol timeline right now uses a strict interpretation of the
paints that we have, but that makes it look as if the paints might not
be loaded completely. Instead, when `hasAllPaintPoints` is `true` (which
means the protocol call has returned and we are done getting paint
points) we should just say that all points for the whole recording are
loaded.