Add paint dots to the protocol timeline
This will show where the actual paint points are, which can be helpful
when trying to understand loading stuff. Maybe this is a bridge too far
in the protocol timeline direction? It's getting pretty tall 👀

I had to statically add classnames to a hidden div, which feels gross,
but I'm not sure of a better way to do this with tailwind.