Motivation and Changes
Soft focus has been a theoretical goal of the front-end team for close
to a month at this point (although it was really just me tinkering and
thinking for the first couple of weeks). And that has produced some
results, but there's still quite a bit in motion, as well as some
unknowns. Technically, all we need to do to enable V1 of soft focus is
not onload regions upon the focus window changing (we still issue the
load command, because if we are already loaded it is a noop, and if the
backend has unloaded a portion of the recording that the user has asked
for, we need to ensure that we reload it). One thing that jumps out as
soon as you enable soft-focus is that we need more specific points to
pass for ranges to backend functions like getHitCounts. I tried just
using getPointNearTime with the focus window times, but that is no good,
because it always returns the beginning of the region, so sub-region
focus windows end up looking like windows of 0 length in point-terms.
Instead, I do a more sophisticated version of what we were doing
previously with timeline seeks:

Any time we get a timestamped point from the backend, we store it in a
sorted array in redux. Then, when the user sets the focus window, we use
that point array to approximate the point that lies near the user's
selected times. This is sometimes not very fine-grained, but here's the
elegant part: if there is stuff that matters around the area that is
being selected (for instance, analysis hits), then we will have stored
fine-grained point correlations already, and we won't be vastly off.
Conversely, if the points we have are really rough approximations, it is
unlikely that they will affect anything anyways, because the user has
not looked at anything of interest around them.

I was prepared for this to be terrible, but in my relatively limited
experience... it's pretty good? I think at the very least it is good
enough to enable it internally and poke around.

Followups:

- We should actually soft-focus analyses, but I will wait on Mark's
rewrite of the analysisManager before making that change.