Only use start and end when soft focus is enabled
There's just a little too much fussiness with focus regions I think for
this to be generally released. A lot of the best benefits will already
be had just by passing ranges to console logs and getHitCounts, and
those will be available for all users. Meanwhile, us internal folks can
mess around with this, figure out when it feels good, when it feels bad,
and roll it out a little slower.