Don't try to create pauses in unloaded regions
We used to have a simpler pause world before focus and turbo replays.
Now, we have to do things like handle pauses that try to get created in
regions that it is not possible to create a pause in (at the moment).
Upon loading, we could do a few things:

- We could not try to pause in unloaded regions, and just go to the end instead
- We could try to load the requested region
- I think the latter is better, but the former is a *lot* easier 😅

Related to #6548 , though it's not a perfect solve.