Don't precache resume targets
From
notion.so/replayio/How-we-use-Pauses-057fc6b63ace4095bd19926cec135596:

Whenever we do succeed in creating a pause, at least for analysis
points, we will then go on to create many more pauses, trying to load
any possible nearby points (sometimes these points are actually not that
nearby, more on that below) that the user might try to go to.

Sometimes, loading those nearby points does not work because they are in
unloaded regions, and again, we should handle that more gracefully than
we do (which right now is blow up with a console error but otherwise do
nothing).

I suspect we should not do this anymore. I think this was probably
useful back when Replay was slower, but right now this seems like it is
just creating havoc on the protocol sometimes, and is not offering a ton
of user benefit. I could be convinced that we should keep it though.