Record failed pause events #7009
This has taken a long time, partially because there was a lot of
complexity around how pauses get created, partially because I went off
on side tangents to simplify things, and partially because I was getting
distracted. Anyways, this is the second part to fixing
https://github.com/RecordReplay/devtools/issues/6548 and
https://github.com/RecordReplay/customer-support/issues/86. Both of
those issues were from devtools trying to create a pause/precache stuff
that was not loaded. And both of them failed two ways:

- First off, we could have known that those calls were not a good idea
  and not made them
- When they failed, we should have told the user there was a failure

As of
https://github.com/RecordReplay/devtools/commit/85e046b8d08bfd7f915d65b08f65a6cfe7d36824
and
https://github.com/RecordReplay/devtools/commit/a303c248836722afde84d135a423a379115d373c
we should not be making requests for pauses that either cannot be
created or will cause an excess of data to load outside of the currently
focused window, and this PR will make sure that there are real error
states in the pause panel when we have failed to pause.