Sort comments by time first, then point
We are apparently sometimes storing comments without `point`s. This is
an important change to be defensive against that case, but it is *also*
an improvement for cases where the `point` and time were far away from
each other because of time <-> point mapping issues.

Fixes https://github.com/RecordReplay/devtools/issues/6865