Don't show broken next and previous buttons
If there are no analysisPoints for a particular line, then we should not
offer to run forward to the next hit, or back to the previous hit. We
could go back to the old way of running an Analysis.findPoints for each
location that this button is held on, but that also seems like not a
great idea, given how performance heavy that can be. I do think it would
be reasonable to do a `Debugger.stepToLocation` method, but we don't
have that yet, so here we are.