Don't hide logs that are in unloaded regions
If the focus mode is being edited, we may show logs that we have (in memory) that are now within an unloaded region. This feels less confusing than if we hid those logs. Assuming the user discards the pending focus region, it doesn't matter– but if they apply it, we're going to immediately load the region again anyway so any kind of inspection they might do would just load like normal after a small delay.