Use separate DB's and don't query so much
Maybe fixes https://github.com/replayio/customer-support/issues/108, see
that issue for more discussion, but the gist is:

Chrome's indexedDB can be slow, and we `await` a bunch of reads from it
when we start up the store. Our theory is that these compete with each
other, and slow down all tabs from loading. I suspect it would be even
*better* if we just wrote to indexedDB way, way less, because as
@gideonred has pointed out - none of the features that we use it to
provide are necessities.