Remove dead code and allow undefined
We had some dead code from removing the `indexing` checks. And, the old
setTimeout game we were playing was specifically for not running
multiple expensive analyses. It does not matter how many times we
request breakpoints, because that gets memoized immediately. It was
unclear to me if we could also pull out the ref there, but I didn't feel
like it was a thread worth tugging on.

I also slightly munged the types for logpoint panel status. I think this
is a little easier to read and will make for a more reasonable default
(`?\?` feels more right to me than `0/0` in the case that we are trying
to present a status and we really just don't have a result)