Show unloaded regions on the breakpoint timeline
Prior to this change, we would only show *unfocused* regions on the
breakpoint timeline, but unloaded regions would appear as if they were
both loaded and focused. Now, we look at the intersection of the loaded
and focused windows.

We still show hits in the unloaded region, which is unfortunate, but I
think the real solution there is not to do filtering, but rather to
adjust the currently focused window when a region gets unloaded.

Before:

<img width="1640" alt="CleanShot 2022-06-09 at 11 25 02@2x" src="https://user-images.githubusercontent.com/5903784/172814328-8b1068b2-13d0-43df-a9c5-d1db698d8e86.png">

After:

<img width="1636" alt="CleanShot 2022-06-09 at 11 56 25@2x" src="https://user-images.githubusercontent.com/5903784/172820506-2c0cabfe-8cb6-46c6-a053-6f76599bd4cc.png">