Architecture exploration: GraphQL (#7099)
This commit explores what a Suspense-cache backed GraphQL system might look like. It adds comments and replies, but the implementation is pretty barebones. GraphQL messaging is done via the Fetch API rather than a library. (Longer term, we may want to explore a technology like Relay, but I wanted to see how far we could get with something simpler.)

The commit also explores the idea of storing pending comments (or replies) in GraphQL as well rather than in-memory (related: #6566). This is something that I think we should do in main – although the exact UX (when to save as private vs public) is still TBD.

Here is a Loom walk through of the changes explored here:
https://www.loom.com/share/2b8b587a738f4913af2371613935da0d