Reload hitCounts when they get cleared due to focus
When the focus region changes, we clear out the hitCounts because they
are no longer reliable. However, the LineHitCounts component was not
refetching them when they got cleared, so they would not update as focus
changed. This seems like the kind of thing which might be addressed well
with one of the new architecture proposals? Until then, this is a slight
modification to the existing hackiness.