If the user is already logged in, redirect
Immediately, without asking them to sign in again. This comes up when
the Replay browser is authing to app.replay.io. It has a query string to
return to some other path, but then it opens Chrome to auth, and Chrome
then reopens Replay browser, but with a new tab on the Home page. The
old tab is still there, pointed at `/login?returnTo=asdklfjkdlsjkl`, but
if you refresh it looks like you *still* need to log in, even though it
clearly says "Rec" and not "Sign in" in the browser. If we land on that
login page, but we have a user, we should just redirect them to the
`returnTo` path.