Avoid creating multiple pauses for an execution point (#7540)
Conceptually, an execution point can have multiple pauses in order to allow forking for analysis with side effects. Practically, I don't think the devtools wants to support this (at least for now). To get a sense of how often execution points end up with more than one pause ID unexpectedly, the devtools now log when this happens to Sentry as an exception.

The TimelineContext has also been updated to remove the requirement for a pauseId so that we can avoid eagerly generating a duplicate. Instead the new Console waits to request the pause after seeking, which means we should be able to reuse a cached pause id. (Locally testing seems to confirm that this is true.)

This commit also updates e2e fixture data and screenshots.