Filter timestamped points by point, not time
We were filtering analysis points by displayed begin and end times. That
does not work. The displayed begin and end times for a focus region
don't necessarily reflect the true begin and end times for that region.
Since we pass a point range for hitCounts, sometimes analysis and hitCounts
disagree. Now, we might have another class of bug: it's theoretically
possible (though hopefully pretty unlikely!) to have "hits" that are
outside of the displayed timeline of a breakpoint. If we see that
happening, we can rethink some of this, but I think it will not happen
*nearly* as often as us miscounting hits because of edge pieces.