Improve object pre-caching for legacy and new architecture interop (#7566)
Address cases of missing pre-cached object preview data due to timing (the connector being further down inside of the tree and subscribing inside of a passive effect).

Also add a fallback to fetch the data (via Suspense) if we ever hit a cache miss. Ideally this wouldn't be necessary but it seems to be in at least some cases (see BAC-2095).