Enable the new Console and Object Inspector by default for all users (#7586)
I think the easiest way to default this to being on for all users is to just rename the preference and change its default to on rather than off.

Note that this PR also updates a lot of our e2e test helper functions and selectors to use data attributes rather than CSS class names. This is not a backwards compatible change. If we revert the new console, we will need to revert (or disable) these tests also.