New Object Inspector UI more closely mirrors Chrome for wrapping and previews (#7622)
New Object Inspector UI more closely mirrors Chrome for wrapping and previews 

When an object is displayed as part of a key+value pair, hide the preview when the object is expanded/inspected. (Note that to keep parity with Chrome, should not hide the preview when the object does not have a key as there would be no way to collapse it again.)

Don't wrap values within the Scopes panel. They should expand off to the right and be hidden (overflow-x: hidden). To inspect those values, either expand the object, hover for a tooltip, or copy/paste. (This is again to maintain parity with Chrome.)