Repaint graphics when we get the right paint #7626
This is a bit of a cludge, but I think it's better than what we have
right now. Basically: if we see paint points out beyond our currently
paused time, we should load the most recent past paint and use that
immediately, rather than waiting for the pause paint to return. If you
compare the loading on the right (prod) with the loading window on the
left (this patch), you should see what I mean.

https://user-images.githubusercontent.com/5903784/186999680-2f8dd54e-7345-40ed-89f3-5b8590582349.mp4