Enable new TypeScript ESlint rules (#7707)
Add TypeScript @typescript-eslint/no-base-to-string to prevent accidentally serializing an object to the string "[Object]". This caused a recent problem with the hit points Suspense cache key (see #7697)

Add @typescript-eslint/no-unnecessary-condition to prevent accidentally using a value (like an object) inside of a conditional– thus getting a false positive. This caused a recent problem with the hit points Suspense cache key (see #7701)