Disable Network and Event Types for Chromium recordings (#7713)
I think that ideally, this type of capabilities metadata would be embedded into a recording by the runtime and passed along to the frontend via the protocol, but in lieu of that this PR adds the concept to the protocol package. Maybe we can build on top of this pattern for other runtime-specific capabilities so that we at least have only a single place to update these things as capabilities change?

Related to FE-673, FE-709, and FE-710.