Set height in requestAnimationFrame for perf
`requestAnimationFrame` fires right after the DOM has just calculated
layout anyways, so reading from dom rects is usually cheaper when done
here, where no recalc is required. If we call `getBoundingClientRect` in
render, we are causing style recalcs, which are expensive.

Before:
https://share.firefox.dev/3dzkSy0

![CleanShot 2022-09-21 at 14 18 33](https://user-images.githubusercontent.com/5903784/191612522-76832ca8-f9e8-4fbe-8f37-528aa849c636.png)

After:
https://share.firefox.dev/3Bxej73

I don't have a screenshot because `getBoundingClientRect` is only in 3 samples so it does not make it on to that list 🙂