Add some more unstable source ID logic
We are changing original source IDs on the backend completely. This is
going to be a good thing in the future, because it means that source IDs
will be reliable. However, in the interim, we have a lot of comments
floating around with original IDs which definitely won't be in the new
format. We had some logic to handle this in one place, but we needed it
in one more place, and I extracted out a function, not so much because
the logic was a clear candidate for abstracting away, but because I
wanted a place to give what we were doing a name and an explanatory
comment (and hopefully try and contain its spread).