Changed preview token hover style
Rather than a background, we now use brightness (or darkness). Some initial amount is applied on hover, and more if the token has been debounced/selected. The second hover stays persistent even if you mouse off of the taken and over the preview popup.

This is different from our past editor, which showed no highlight at all (only the preview tooltip tail, which we also show).