Add the ability to cache scan data
And change what it means for regions to be "loaded". We have several
different things we track on the the region level:

- Loading: are we ever going to try and load this region given the
  current focus window?
- Indexed: have we processing the manifests which produce scan data for
  this region, which allows things like hitCounts to work
- Loaded: Have we done a run-to-point and fork at the beginning of this
  region yet or not?

Ironically, for the colloquial "loading" case (can I interact with the
recording) we don't really care about the `loaded` state as defined
above. If we have not RTP and forked in a region, then doing things like
pause operations/analyses will be slow there, but we can totally get
started queueing up those requests, they'll just take a little longer to
respond. By not including that in the loading percentage we can get
people interacting with editor, hitcounts, paints, etc. faster.