Load console messages earlier
and use `indexed` rather than `loading` for checking whether we should
consider a part of the recording ready to interact with.

From a comment on the PR:

> We have two different methods for fetching console messages:
findMessages and findMessagesInRange. Now, I've gone on this rant
probably a dozen times at this point, but the reason that this one basic
data resource has multiple methods for fetching exactly the same kind of
information is because it has an overflow flag, so requesting console
messages in a particular region or set of regions might allow you to
bypass that flag. I wish we did not have this, but I'm also not sure of
the best way to change it., and unfortunately, these commands have
different requirements/expectations because of how they work.
>
> findMessages is much closer to all of the other basic data resources: it
does not interact with loaded or indexed regions at all. It's always
available, and always returns all data for the whole recording,
regardless of what is currently loaded. For instance, we do not wait for
the whole recording to be loaded before calling findSources, nor do we
expect that findSources will only return sources that were used or first
downloaded in the currently loaded regions.
>
> Because findMessagesInRange is designed to get around the overflow flag,
it has to go directly to the running linker and ask it to run through
particular part of the recording and re-collect the console messages.
This requires a running linker in those regions, and so it requires that
those regions be loaded. I think we should either make this endpoint
return the actual window for which it ran with the results, or have it
error if it can't run through the requested window, rather than what it
does now (silently return a different window) but that's a slightly
tangential topic.
>
> Anyways, because of the way findMessages works, we can issue that
request as soon as the Replay is loaded, and in most cases we will never
need to request again, because there will be fewer than 1,000 console
messages.