Be defensive about URL parsing
I just hit this on a recording where I manually changed the URL to
`foobar`. That's probably not a common case, but we *do* have an API
where users can set the recording metadata to whatever they want. Given
that, we should probably be careful when parsing that info to not throw.