Don't wrap parens around expressions sent to Pause.evaluateInGlobal (#8364)
We were wrapping expressions in parens (e.g. a + 1 was modified to (a + 1) before being sent to Pause.evaluateInGlobal). This operation could cause a valid expression to become invalid.

I'm not sure why we were doing this, but it doesn't seem to be necessary based on some limited testing, so this commit removes the parens.