Load box model data as soon as a node is selected
The original FF markup logic waited for the "Inspector" panel (ie
the HTML "Elements" panel) to be selected before it would fetch
box model data.

That logic was preserved when we migrated the rest of the "Elements"
handling over to be fully controlled by Redux.

From there, the Redux logic was checking to see if we have markup
data for this selected node.

Thinking about it further, I'm not sure there's really a good reason
to wait. API calls to fetch box model data aren't heavy, and we
cache the results. The more important thing is to make sure that
we keep the displayed data in sync with the selected node.

So, we might as well just fetch the box model values and show them
as soon as a node is selected.

Also applied the same logic to the "Rules" panel fetching.