Refactor sidebar button logic and fix collapsed persistence
The state of the sidebar panel was originally completely kept in
Redux, in a `startPanelCollapsed` field. At some point, with the
migration to the new "resizeable panels" library, we stopped using
that field to drive the UI.  However, it was still getting persisted
and read in one place: `<ToolbarButton>`

Somehow we ended up with two different sets of checks inside the
the click handler in `ToolbarButton`. The mismatched logic led to
the click handler thinking that the panel was already open when
it was actually closed, so sometimes a click on a button wouldn't
do anything.

I've refactored the logic to pull the toggle handling up into
`<Toolbar>`, removed the Redux `start/endPanelCollapsed` fields
and all the persistence related to those, and added persistence of
the component state `sidePanelCollapsed` field via `useLocalStorage`.