Restructure Source Viewer to not block on hit counts or positions
The `<SourceList>` component needs 3 major pieces of data: source
contents, hit counts, and breakable positions. It had been fetching
all of those using the `getDataSuspense` methods. However, those
"block" inside of the component, and won't continue until the data
has resolved.

We saw a case where the hit counts request never resolved, and that
kept `<SourceList>` from actually showing the selected file's text.
But, hit counts shouldn't be critical-path data for rendering the
_text_ of a file. We should always show the text as soon as it's
available, and show hit counts and breakable positions in a
non-blocking way.

Conveniently, we already had a `createUseGetValue` hook factory,
which fetches via one of our "getDataAsync" methods and sets state.

The hit counts cache is custom and _not_ a generic cache, so I had
to restructure the logic to expose all the same methods as a generic
cache would.