Remove unprocessedRegions usage
The whole `ensureProcessed` command is probably going to either get
deleted or change considerably in the near future. It seems like
basically all of this is not really powering any important functionality
at the moment. The SourceMap visualizer has a comment about showing a
loading bar, but actually removing this has no affect on the source map
viewer loading screen.