Dont request all possible breakpoints
As of 90b36c591 we no longer need to worry about requesting all possible
breakpoints for correspondingSourceIds.

Also, I think there was a bug where we would request possible
breakpoints in an n*2 loop, which we probably did not notice because the
number of corresponding sources is normally low, but for Cypress replays
it's commonly *many* (dozens or more).