Use jest-mock-extendded to generate ReplayClientInterface mocks (#9101)
When working on #9098, I noticed that the TypeScript for our mock ReplayClientInterface methods were all any – which isn't great, but also that mock was getting pretty unwieldy to maintain.

To address both of these issues, I added the jest-mock-extended package to generate this mock with proper types. Also found and fixed a few minor type errors along the way.