Write Delta snapshot files to unique paths per test (#9221)
Initially screenshots were being saved to a directory that included the test file name and the description, e.g. `packages/replay-next/playwright/snapshots/tests/console/should-be-filterable-on-complex-content.ts-snapshots/dark/…`

As of PR #8407, screenshots for CI tests we being written to a flat directory, `packages/replay-next/playwright/visuals/dark/…`

This made a race condition for tests that happened to have screenshots with the same names. This PR restores something more like the original behavior. (Note that it will generate _a lot_ of Delta changes.)