Improve Source preview popup error messages (#9719)
When the user hovers over a value in the source editor that can't be inspected, we used to just show a hard-coded message. Andarist shared feedback that it was confusing the Scopes panel showed something different for these same variables ("unavailable"). He suggested we show the same thing in both locations, but after thinking about this more– I think it's better if we take a different approach:
  1. Show the specific error reported by the Protocol. (It will have clearer text than just showing "(unavailable)")
  2. Style the popup differently when we're displaying a meta error (to differentiate when we're inspecting an `Error` object)

I don't think it makes sense to change what we show in the Scopes panel to match this, since it's more of a passive panel, but this seems subjective and I'm open to discussion.