Render log point panel loader before suspending (#9947)
The trick here was to render the log point panel loader before the update that suspended, as that update also included other suspending components (like LoggablesContext) that sometimes would prevent the update from committing the loader. This feels like an awkward React+Suspense edge case but I'm not sure how to work around it in another way.