Add client ID header for GraphQL requests
The Problem
---
When debugging GraphQL issues, especially E2E failures, it can get
really confusing trying to figure out which backend GraphQL logs
correspond to which client.

Proposed Solution
---
By generating a UUID when we create the
client and attaching that to a header which the backend will log we can
make it a lot easier to align backend log lines with recordings of
devtools to debug test failures.