Remove "basic processing" estimation and associated Redux state (#10227)
This feature was a hacky workaround for the problem of loading being "stuck" (DES-82). We now have a much better solution in place for this– explicitly wait on the backend to complete processing before loading DevTools UI (FE-2170).

This feature was disabled by default as an advanced setting, so deleting it should have minimal user impact.

Removing this code allows us to also remove the redundant timeline.paints and timeline.points values from Redux, which has the added benefit of unblocking some planned refactoring work.