Handle "not-found" and "error" responses from component inspection (#10234)
This was an oversight in my previous implementation. The React DevTools has a few possible message types:
```ts
export type InspectedElementResponseType =
  | 'error'
  | 'full-data'
  | 'hydrated-path'
  | 'no-change'
  | 'not-found';
```

Some of these are not relevant in Relay's usage (like "hydrated-path" for example, since we fetch values with Replay's protocol instead) but I should have been handling the "not-found" case, and I guess to be complete– the "error" case as well. This commit adds handling for those two, along with a user-facing error message.

![Screenshot 2024-01-26 at 10 34 34 AM](https://github.com/replayio/devtools/assets/29597/2fd63136-13eb-43d0-a603-ae3a1245458f)