Suspense caches for paint (points) and recorded mouse events (#10230)
Begin moving state out of the protocol/graphics package and into our typical suspense-caches. Eventually we've move the imperative logic out of that package as well.

This commit also removes the internal ReplayClient.getSessionIdThrows() method. It seems dangerous in that it can lead to an error being thrown if an API is called too early. These APIs are async anyway so it doesn't seem to be worth the risk to avoid the overhead of an extra promise tick.

This is an incremental step toward some larger planned refactoring work.