More explict type guard for recordings without Paints (#10244)
PR #10230 broke playback for Node recordings because it didn't properly handle the case of array out-of-bounds. Unfortunately I wish TypeScript was stricter about that in our project, but for now I've updated the PaintsCache to return an explicit null value/type for Node recordings, to force us to be more explicit with our handling of the no-paints case.