Removed time-is-beyond-known-paints check (#10263)
Paints are expected to load in very quickly. In fact, because of this, we recently removed the streaming behavior for them. Given that, the old check didn't make sense anymore.

I verified that after removing this check, trying to "play" after the last paint works like it used to (which is to say, the timeline marker moves forward and graphics stay the same).