Remove "protocol_chromiumRepaints" experimental setting (#10318)
The concept of whether a runtime "supported repaint" was originally added in back in 2022 (#7713). At that time, the Chromium platform did not yet support paints (or it was very unstable) so we added an experimental setting for it. Then later on we flipped that setting on by default and forgot about it.

This doesn't make sense to track as an experimental setting anymore, as that feature is stable. What's more, it recently caused @jcmorrow some confusion when he accidentally turned it off and then found that the video panel was missing for Chromium recordings.