[RUN-3292] separately targetable steps for e2e on linux-x86_64 and macos-arm64 (#10392)
Support running the e2e tests on both platforms, since we've had issues affecting only macos-arm64 in the past.

In an effort to not wait on both chromium builds before starting the tests, make each step targetable using env vars.  The corresponding PR in chromium (https://github.com/replayio/chromium/pull/1142) waits on a single build and then triggers this pipeline with the proper env vars.  This could also be done with multiple buildkite pipelines and/or multiple .yml files, but this seemed the best choice to keep things together.