Pass testScope through GraphQL Subscription Request
In order to properly scope comments we need to have the testScope
available on the websocket request. Headers on websocket requests are
weird, you are not really supposed to use them, so we pass it through
the same way we do the auth details.

This is a noop on the backend right now, but soon it will be used.