Improve repaint/graphics e2e test stability (#10419)
I noticed some repaint e2e test failures recently. They aren't common, but when they occur, it seems like our logic of waiting for a screenshot to change is not robust enough to handle a delay in loading the initial screenshot, which can cause subsequent comparisons to fail.

To reduce the likelihood of this, I've updated our test helpers to explicitly wait for the current screenshot to finish loading before returning the graphics content. I think this should improve the stability of the various repaint related tests.