Improve HTMLElement type detection and rendering (#10468)
It's been over a year since this code was written, so my recollection is a bit foggy. I believe the reasons I specified a preview level of `"none"` initially were:
1. It's less data to fetch in a hot part of our code
2. Most value types don't require a preview; only HTML elements do

That being said, the result of this approach was that HTML elements _may_ render incorrectly depending on the context they were fetched in. (If they are a nested value, they may not have preview information.)

We could change the initial request to always fetch some preview information (`"canOverflow"`) but I'm not sure how much of a performance impact that may have, so instead I opted to only fallback to fetching preview information when we've ruled out all other possible value types.